+++ affinda/models/_models.py
@@ -10219,6 +10219,8 @@ class JobDescriptionSearchParameters(msrest.serialization.Model):
     :vartype management_level_required: bool
     :ivar management_level_weight:
     :vartype management_level_weight: float
+    :ivar custom_data:
+    :vartype custom_data: list[~affinda.models.SearchParametersCustomData]
     """
 
     _validation = {
@@ -10265,6 +10267,7 @@ class JobDescriptionSearchParameters(msrest.serialization.Model):
         "management_level": {"key": "managementLevel", "type": "str"},
         "management_level_required": {"key": "managementLevelRequired", "type": "bool"},
         "management_level_weight": {"key": "managementLevelWeight", "type": "float"},
+        "custom_data": {"key": "customData", "type": "[SearchParametersCustomData]"},
     }
 
     def __init__(self, **kwargs):
@@ -10327,6 +10330,8 @@ class JobDescriptionSearchParameters(msrest.serialization.Model):
         :paramtype management_level_required: bool
         :keyword management_level_weight:
         :paramtype management_level_weight: float
+        :keyword custom_data:
+        :paramtype custom_data: list[~affinda.models.SearchParametersCustomData]
         """
         super(JobDescriptionSearchParameters, self).__init__(**kwargs)
         self.indices = kwargs["indices"]
@@ -10358,6 +10363,7 @@ class JobDescriptionSearchParameters(msrest.serialization.Model):
         self.management_level = kwargs.get("management_level", None)
         self.management_level_required = kwargs.get("management_level_required", None)
         self.management_level_weight = kwargs.get("management_level_weight", None)
+        self.custom_data = kwargs.get("custom_data", None)
 
 
 class JobDescriptionSearchResult(msrest.serialization.Model):
@@ -10391,6 +10397,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
     :vartype search_expression: ~affinda.models.SearchExpressionSearchScoreComponent
     :ivar organization_name: Required.
     :vartype organization_name: str
+    :ivar custom_data: Required. Dictionary of
+     <components·nqbw24·schemas·customdatasearchscorecomponent·additionalproperties>.
+    :vartype custom_data: dict[str,
+     ~affinda.models.ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties]
     """
 
     _validation = {
@@ -10406,6 +10416,7 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         "education": {"required": True},
         "search_expression": {"required": True},
         "organization_name": {"required": True},
+        "custom_data": {"required": True},
     }
 
     _attribute_map = {
@@ -10431,6 +10442,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
             "type": "SearchExpressionSearchScoreComponent",
         },
         "organization_name": {"key": "organizationName", "type": "str"},
+        "custom_data": {
+            "key": "customData",
+            "type": "{ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties}",
+        },
     }
 
     def __init__(self, **kwargs):
@@ -10461,6 +10476,10 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         :paramtype search_expression: ~affinda.models.SearchExpressionSearchScoreComponent
         :keyword organization_name: Required.
         :paramtype organization_name: str
+        :keyword custom_data: Required. Dictionary of
+         <components·nqbw24·schemas·customdatasearchscorecomponent·additionalproperties>.
+        :paramtype custom_data: dict[str,
+         ~affinda.models.ComponentsNqbw24SchemasCustomdatasearchscorecomponentAdditionalproperties]
         """
         super(JobDescriptionSearchResult, self).__init__(**kwargs)
         self.identifier = kwargs["identifier"]
@@ -10476,6 +10495,7 @@ class JobDescriptionSearchResult(msrest.serialization.Model):
         self.occupation_group = kwargs.get("occupation_group", None)
         self.search_expression = kwargs["search_expression"]
         self.organization_name = kwargs["organization_name"]
+        self.custom_data = kwargs["custom_data"]
 
 
 class JobTitleAnnotation(Annotation):
@@ -11955,7 +11975,7 @@ class Paths1TvfqeiV3IndexPostResponses201ContentApplicationJsonSchema(msrest.ser
     :ivar name:
     :vartype name: str
     :ivar document_type: Known values are: "resumes", "job_descriptions".
-    :vartype document_type: str or ~affinda.models.Enum20
+    :vartype document_type: str or ~affinda.models.Enum21
     """
 
     _attribute_map = {
@@ -11968,7 +11988,7 @@ class Paths1TvfqeiV3IndexPostResponses201ContentApplicationJsonSchema(msrest.ser
         :keyword name:
         :paramtype name: str
         :keyword document_type: Known values are: "resumes", "job_descriptions".
-        :paramtype document_type: str or ~affinda.models.Enum20
+        :paramtype document_type: str or ~affinda.models.Enum21
         """
         super(Paths1TvfqeiV3IndexPostResponses201ContentApplicationJsonSchema, self).__init__(
             **kwargs
@@ -14835,14 +14855,14 @@ class ResumeSearchParameters(msrest.serialization.Model):
         self.custom_data = kwargs.get("custom_data", None)
 
 
-class ResumeSearchParametersCustomData(msrest.serialization.Model):
-    """ResumeSearchParametersCustomData.
+class SearchParametersCustomData(msrest.serialization.Model):
+    """SearchParametersCustomData.
 
     All required parameters must be populated in order to send to Azure.
 
     :ivar filter_type: Required. Data points of "text" type support only "equals" filterType,
      others support both "equals" and "range". Known values are: "equals", "range".
-    :vartype filter_type: str or ~affinda.models.ResumeSearchParametersCustomDataFilterType
+    :vartype filter_type: str or ~affinda.models.SearchParametersCustomDataFilterType
     :ivar data_point: Required. The data point's slug.
     :vartype data_point: str
     :ivar query: Required. "equals" searches require the "value" key inside the query, and "range"
@@ -14873,7 +14893,65 @@ class ResumeSearchParametersCustomData(msrest.serialization.Model):
         """
         :keyword filter_type: Required. Data points of "text" type support only "equals" filterType,
          others support both "equals" and "range". Known values are: "equals", "range".
-        :paramtype filter_type: str or ~affinda.models.ResumeSearchParametersCustomDataFilterType
+        :paramtype filter_type: str or ~affinda.models.SearchParametersCustomDataFilterType
+        :keyword data_point: Required. The data point's slug.
+        :paramtype data_point: str
+        :keyword query: Required. "equals" searches require the "value" key inside the query, and
+         "range" searches require at least one of "gte" (greater than or equal) and "lte" (less than or
+         equal).
+        :paramtype query: any
+        :keyword required:
+        :paramtype required: bool
+        :keyword weight:
+        :paramtype weight: float
+        """
+        super(SearchParametersCustomData, self).__init__(**kwargs)
+        self.filter_type = kwargs["filter_type"]
+        self.data_point = kwargs["data_point"]
+        self.query = kwargs["query"]
+        self.required = kwargs.get("required", None)
+        self.weight = kwargs.get("weight", None)
+
+
+class ResumeSearchParametersCustomData(SearchParametersCustomData):
+    """ResumeSearchParametersCustomData.
+
+    All required parameters must be populated in order to send to Azure.
+
+    :ivar filter_type: Required. Data points of "text" type support only "equals" filterType,
+     others support both "equals" and "range". Known values are: "equals", "range".
+    :vartype filter_type: str or ~affinda.models.SearchParametersCustomDataFilterType
+    :ivar data_point: Required. The data point's slug.
+    :vartype data_point: str
+    :ivar query: Required. "equals" searches require the "value" key inside the query, and "range"
+     searches require at least one of "gte" (greater than or equal) and "lte" (less than or equal).
+    :vartype query: any
+    :ivar required:
+    :vartype required: bool
+    :ivar weight:
+    :vartype weight: float
+    """
+
+    _validation = {
+        "filter_type": {"required": True},
+        "data_point": {"required": True},
+        "query": {"required": True},
+        "weight": {"maximum": 1, "minimum": 0},
+    }
+
+    _attribute_map = {
+        "filter_type": {"key": "filterType", "type": "str"},
+        "data_point": {"key": "dataPoint", "type": "str"},
+        "query": {"key": "query", "type": "object"},
+        "required": {"key": "required", "type": "bool"},
+        "weight": {"key": "weight", "type": "float"},
+    }
+
+    def __init__(self, **kwargs):
+        """
+        :keyword filter_type: Required. Data points of "text" type support only "equals" filterType,
+         others support both "equals" and "range". Known values are: "equals", "range".
+        :paramtype filter_type: str or ~affinda.models.SearchParametersCustomDataFilterType
         :keyword data_point: Required. The data point's slug.
         :paramtype data_point: str
         :keyword query: Required. "equals" searches require the "value" key inside the query, and
@@ -14886,11 +14964,6 @@ class ResumeSearchParametersCustomData(msrest.serialization.Model):
         :paramtype weight: float
         """
         super(ResumeSearchParametersCustomData, self).__init__(**kwargs)
-        self.filter_type = kwargs["filter_type"]
-        self.data_point = kwargs["data_point"]
-        self.query = kwargs["query"]
-        self.required = kwargs.get("required", None)
-        self.weight = kwargs.get("weight", None)
 
 
 class ResumeSearchParametersLocation(msrest.serialization.Model):